refactor files, add doc comments
