+++ internal/overlord/logstate/forward.go
@@ -1,14 +1,7 @@
 package logstate
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"net"
-	"net/http"
-	"net/url"
-	"strconv"
 	"sync"
 	"time"
 
@@ -19,21 +12,29 @@ var maxLogBytes int = 100 * 1024
 
 const canonicalPrivEnterpriseNum = 28978
 
+// LogBackend defines an interface to facilitate support for different log destination "types"
+// (e.g. syslog, loki, etc.).
 type LogBackend interface {
 	Send(*LogMessage) error
 	Close() error
 }
 
+// LogMessage represents service-originated log content with metadata relevant to log forwarding.
 type LogMessage struct {
 	Service   string
 	Message   []byte
 	Timestamp time.Time
 }
 
+// Size returns the size (i.e. in bytes) of the message - useful for e.g. limiting memory usage for
+// buffered logs.
 func (l *LogMessage) Size() int {
 	return len(l.Message)
 }
 
+// LogForwarder is a simple writer that is used to intercept (stdout+stderr) output from running services
+// and annotate it with relevant metadata (e.g. service name) before conveying it to a particular
+// log destination.
 type LogForwarder struct {
 	service     string
 	destination *LogDestination
@@ -49,14 +50,22 @@ func (l *LogForwarder) Write(p []byte) (int, error) {
 	// If so, how?
 	err := l.destination.Send(&LogMessage{Message: data, Service: l.service, Timestamp: time.Now()})
 	if err != nil {
-		logger.Noticef("log destination write failed: %v", err)
+		logger.Noticef("failed to transmit service %q logs to destination %q: %v", l.service, l.destination.name, err)
 		return 0, err
 	}
 	return len(p), nil
 }
 
+// LogDestination manages fowarding log content to some destination.  It can be configured to send
+// logs to any sort of underlying destination (e.g. syslog, loki, etc.) by priming it with an
+// appropriate backend.  Messages accumulate in a fixed-size internal buffer and are concurrently
+// pushed/sent to the underlying backend.  If the backend is unable to keep up and the buffer fills
+// up, logs are discarded incrementally as necessary in FIFO order.  LogDestination is safe to use
+// concurrently.  Close should be called when the destination will no longer be used to clean up
+// internal goroutines, etc.
 type LogDestination struct {
 	mu      sync.Mutex
+	name    string
 	buf     *LogBuffer
 	notify  chan bool
 	backend LogBackend
@@ -64,9 +73,10 @@ type LogDestination struct {
 	closed  bool
 }
 
-func NewLogDestination(backend LogBackend) *LogDestination {
+func NewLogDestination(name string, backend LogBackend) *LogDestination {
 	buf := NewLogBuffer(maxLogBytes)
 	c := &LogDestination{
+		name:    name,
 		backend: backend,
 		buf:     buf,
 		notify:  make(chan bool, 1),
@@ -78,6 +88,7 @@ func NewLogDestination(backend LogBackend) *LogDestination {
 	return c
 }
 
+// SetBackend updates the backend to which log messages are sent.  This method is concurrency-safe.
 func (c *LogDestination) SetBackend(b LogBackend) error {
 	if c.closed {
 		return fmt.Errorf("cannot modify a closed destination")
@@ -91,6 +102,7 @@ func (c *LogDestination) SetBackend(b LogBackend) error {
 	return nil
 }
 
+// Send conveys msg to the underlying backend.
 func (c *LogDestination) Send(msg *LogMessage) error {
 	if c.closed {
 		return fmt.Errorf("cannot send messages to a closed destination")
@@ -98,6 +110,9 @@ func (c *LogDestination) Send(msg *LogMessage) error {
 	return c.buf.Put(msg)
 }
 
+// Close marks the destination as no longer available for use.  Further calls to Send will return
+// an error.  Internal resources/goroutines will be cleaned up.  Any unsent buffered logs will
+// *not* be forwarded to the backend.
 func (c *LogDestination) Close() {
 	c.mu.Lock()
 	defer c.mu.Unlock()
@@ -119,238 +134,74 @@ func (c *LogDestination) run() {
 				c.mu.Unlock()
 			}
 		case <-c.done:
+			// TODO: should we try to drain the buffer first?
 			return
 		}
 	}
 }
 
-type LokiBackend struct {
-	address string
+type LogBuffer struct {
+	mu       sync.Mutex
+	capacity int
+	buf      []*LogMessage
+	currSize int
+	notify   []chan<- bool
 }
 
-func NewLokiBackend(address string) (*LokiBackend, error) {
-	u, err := url.Parse(address)
-	if err != nil || u.Host == "" {
-		u, err = url.Parse("//" + address)
+func NewLogBuffer(capacity int) *LogBuffer {
+	return &LogBuffer{
+		capacity: capacity,
 	}
-	if err != nil {
-		return nil, fmt.Errorf("invalid loki server address: %v", err)
-	} else if u.Scheme != "" && u.Scheme != "http" {
-		return nil, fmt.Errorf("unsupported loki address scheme '%v'", u.Scheme)
-	} else if u.RequestURI() != "" {
-		return nil, fmt.Errorf("invalid loki address: extraneous path %q", u.RequestURI())
-	}
-
-	// check for and set loki defaults if ommitted from address
-	if u.Scheme == "" {
-		u.Scheme = "http"
-	}
-	if u.Port() == "" {
-		u.Host += ":3100"
-	}
-
-	b := &LokiBackend{address: address}
-	return b, nil
 }
 
-func (b *LokiBackend) Close() error { return nil }
-
-type lokiMessageStream struct {
-	Stream map[string]string `json:"stream"`
-	Values [][2]string       `json:"values"`
+func (b *LogBuffer) Notify(ch chan<- bool) {
+	b.notify = append(b.notify, ch)
 }
 
-type lokiMessage struct {
-	Streams []lokiMessageStream `json:"streams"`
-}
+func (b *LogBuffer) Put(m *LogMessage) error {
+	b.mu.Lock()
+	defer b.mu.Unlock()
 
-func (b *LokiBackend) Send(m *LogMessage) error {
-	timestamp := strconv.FormatInt(m.Timestamp.UnixNano(), 10)
-	data, err := json.Marshal(lokiMessage{
-		Streams: []lokiMessageStream{
-			lokiMessageStream{
-				Values: [][2]string{{timestamp, string(m.Message)}},
-			},
-		}})
-	if err != nil {
-		logger.Noticef("failed to build loki message: %v", err)
+	if m.Size() > b.capacity {
+		return fmt.Errorf("LogBuffer.Put failed: capacity %v cannot fit object of size %v", b.capacity, m.Size())
 	}
 
-	buf := bytes.NewBuffer(data)
-	addr := "http://" + b.address + "/loki/api/v1/push"
-	logger.Noticef("loki backend is sending message (addr=%v):\n%v", addr, buf.String())
-	r, err := http.NewRequest("POST", addr, buf)
-	if err != nil {
-		logger.Noticef("loki send failed: %v", err)
-		return err
-	}
-	r.Header.Add("Content-Type", "application/json")
-	c := &http.Client{}
-	resp, err := c.Do(r)
-	if err != nil {
-		logger.Noticef("loki send failed: %v", err)
-	}
-	defer resp.Body.Close()
-	data, _ = ioutil.ReadAll(resp.Body)
-	if resp.StatusCode != http.StatusOK {
-		logger.Noticef("%s", data)
-	}
-	return err
-}
-
-// SyslogBackend takes writes and formats them according to RFC5424.  The formatted syslog messages
-// are then forwarded to the specified underlying destination io.Writer.  SyslogWriter is safe for
-// concurrent writes and use.
-type SyslogBackend struct {
-	mu             sync.RWMutex
-	version        int
-	host           string
-	msgid          string
-	priority       int
-	pid            string
-	structuredData string
-
-	conn          net.Conn
-	address       *url.URL
-	waitReconnect time.Duration
-	closed        bool
-}
-
-// NewSyslogBackend creates a writer forwarding writes as syslog messages to dst.  The forwarded
-// messages will have app as the application name.  Other message parameters are set using
-// reasonable defaults or the RFC5424 nil value "-".
-func NewSyslogBackend(addr string) (*SyslogBackend, error) {
-	u, err := url.Parse(addr)
-	if err != nil || u.Host == "" {
-		u, err = url.Parse("//" + addr)
-	}
-	if err != nil {
-		return nil, err
-	} else if u.Scheme != "tcp" && u.Scheme != "udp" && u.Scheme != "" {
-		return nil, fmt.Errorf("invalid syslog server address scheme %q", u.Scheme)
-	}
-	if u.Scheme == "" {
-		u.Scheme = "tcp"
+	available := b.capacity - b.currSize
+	need := m.Size()
+	if available < need {
+		b.discard(need - available)
 	}
 
-	return &SyslogBackend{
-		version:        1,
-		pid:            "-",
-		host:           "-",
-		msgid:          "-",
-		priority:       1*8 + 6, // for facility=user-msg severity=informational. See RFC 5424 6.2.1 for available codes.
-		structuredData: syslogStructuredData("pebble", canonicalPrivEnterpriseNum, nil),
+	b.currSize += m.Size()
+	b.buf = append(b.buf, m)
 
-		address: u,
-	}, nil
-}
-
-// syslogStructuredData formats the given labels into a structured data section for a syslog message
-// according to RFC5424 section 6.
-func syslogStructuredData(name string, enterpriseNum int, labels map[string]string) string {
-	if len(labels) == 0 {
-		return "-"
-	}
-	var buf bytes.Buffer
-	fmt.Fprintf(&buf, "[%s@%d", name, enterpriseNum)
-	for key, value := range labels {
-		fmt.Fprintf(&buf, " %s=\"", key)
-		// escape the value according to RFC5424 6.3.3
-		for i := 0; i < len(value); i++ {
-			// don't use "for _, c := range value" as we don't want runes
-			c := value[i]
-			if c == '"' || c == '\\' || c == ']' {
-				buf.WriteByte('\\')
-			}
-			buf.WriteByte(c)
+	for _, ch := range b.notify {
+		select {
+		case ch <- true:
+		default:
 		}
-		buf.WriteByte('"')
 	}
-	buf.WriteByte(']')
-	return buf.String()
-}
-
-func (s *SyslogBackend) Send(m *LogMessage) error {
-	err := s.ensureConnected()
-	if err != nil {
-		return err
-	}
-
-	s.mu.RLock()
-	_, err = s.conn.Write(s.buildMsg(m))
-	s.mu.RUnlock()
-
-	if err != nil {
-		// The connection might be bad. Close and reset it for later reconnection attempt(s).
-		s.mu.Lock()
-		s.conn.Close()
-		s.mu.Unlock()
-	}
-	return err
-}
-
-func (s *SyslogBackend) buildMsg(m *LogMessage) []byte {
-	// format defined by RFC 5424
-	timestamp := m.Timestamp.Format(time.RFC3339)
-	msg := fmt.Sprintf("<%d>%d %s %s %s %s %s %s %s",
-		s.priority, s.version, timestamp, s.host, m.Service, s.pid, s.msgid, s.structuredData, m.Message)
-
-	// Octet framing as per RFC 5425.
-	framed := fmt.Sprintf("%d %s", len(msg), msg)
-	return []byte(framed)
-}
-
-func (s *SyslogBackend) Close() error {
-	s.closed = true
-	func() {
-		s.mu.RLock()
-		defer s.mu.RUnlock()
-		if s.conn != nil {
-			s.conn.SetDeadline(time.Now().Add(-1 * time.Second))
-		}
-	}()
-
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.conn == nil {
-		return nil
-	}
-	err := s.conn.Close()
-	s.conn = nil
-	return err
-}
-
-func (s *SyslogBackend) ensureConnected() error {
-	if s.conn != nil {
-		return nil
-	} else if s.closed {
-		return fmt.Errorf("write to closed SyslogBackend")
-	}
-
-	if s.waitReconnect > 0 {
-		time.Sleep(s.waitReconnect)
-	}
-
-	conn, err := net.Dial(s.address.Scheme, s.address.Host)
-
-	if err != nil {
-		// start an exponential backoff for reconnection attempts
-		if s.waitReconnect == 0 {
-			s.waitReconnect = 100 * time.Millisecond
-		}
-		newWait := 2 * s.waitReconnect
-		if newWait > 10*time.Second {
-			newWait = 10 * time.Second
-		}
-		s.waitReconnect = newWait
-		return err
-	}
-
-	if s.conn != nil {
-		s.conn.Close()
-	}
-
-	s.waitReconnect = 0 // reset backoff
-	s.conn = conn
 	return nil
+}
+
+func (b *LogBuffer) GetAll() []*LogMessage {
+	b.mu.Lock()
+	defer b.mu.Unlock()
+	buf := b.buf
+	b.buf = b.buf[:0]
+	b.currSize = 0
+	return buf
+}
+
+func (b *LogBuffer) discard(n int) {
+	freed := 0
+	for i, m := range b.buf {
+		freed += m.Size()
+		if freed >= n {
+			b.buf = b.buf[i+1:]
+			b.currSize -= freed
+			return
+		}
+	}
+	panic("failed to free up enough enough space")
 }
